refactor(CRUD Policy):
Moved Policy and data manipulation to policy and request

+++ app/Http/Controllers/ProjectTodoController.php
@@ -55,26 +55,17 @@ class ProjectTodoController extends Controller
      */
     public function store($project_id, StoreTodoRequest $request): RedirectResponse
     {
+        $user = User::find(auth()->user()->id);
+        $this->authorize('create', [Todo::class, $user]);
+
         $validatedData = $request->validated();
 
-        if(isset($validatedData['due_start']))
-            $validatedData['due_start'] = strtotime($validatedData['due_start']);
-
-        if (isset($validatedData['due_end']))
-            $validatedData['due_end'] = strtotime($validatedData['due_end']);
-        elseif (isset($validatedData['due_start']))
-            $validatedData['due_end'] = $validatedData['due_start'];
-//
-        $todo = new Todo($validatedData);
-
-        $user = User::find(auth()->user()->id);
         $project = $user->projects->find($project_id);
         // Add the Todo to the Project
-        $project->todos()->save($todo);
+        $project->todos()->save(new Todo($validatedData));
 
         return redirect()->route('project.todo.index', $project_id)
             ->with('success', 'Todo created successfully.');
-//            ->setStatusCode(201);
     }
 
     /**
@@ -101,12 +92,8 @@ class ProjectTodoController extends Controller
         $projects = $user->projects;
         $project = $projects->find($project_id);
 
-        if (!$project || $project->user->id !== auth()->user()->id || $todo->user()[0]->id !== auth()->user()->id)
-            return back()->with('error', 'Project/Todo not found');
+        $this->authorize('update', [Todo::class, $project, $todo]);
 
-        // Check if the given todo is in the given project (Reverse find with todo's project_id)
-        if ($todo->project->id !== $project_id)
-            return back()->with('error', 'Todo not found in the given project');
 
         return view('todo.edit', compact('project', 'todo'));
     }
@@ -114,39 +101,17 @@ class ProjectTodoController extends Controller
     /**
      * Update Todo in storage based on the given project
      */
-    public function update($project_id, Request $request, Todo $todo)
+    public function update($project_id, UpdateTodoRequest $request, Todo $todo)
     {
-        if ($todo->project->id !== $project_id) {
-            return back()->with('error', 'Todo not found in the given project');
-        }
+        $project = auth()->user()->projects->find($project_id);
 
-        $data = Request::only(['title', 'description', 'due_start', 'due_end', 'completed_at']);
+        $this->authorize('update', [Todo::class, $project, $todo]);
 
-        if (Request::filled('completed_at')) {
-            $todo->completed_at = Request::input('completed_at') === 'on' ? strtotime(now($this->timezone)) : null;
-            $todo->save();
-            return back()->with('success', 'Todo updated successfully');
-        } else {
-            // If 'completed_at' is not provided, toggle its value (only if the request is empty)
-            if (empty($data))
-                $todo->completed_at = $todo->completed_at ? null : strtotime(now($this->timezone));
-            else
-                // Continue to update other fields
-                unset($data['completed_at']);
-        }
+        // Update other fields
+        $todo->fill($request->validated());
 
-        if (Request::filled('due_start')) {
-            $data['due_start'] = strtotime(Request::input('due_start'));
-        }
 
-        if (Request::filled('due_end')) {
-            $data['due_end'] = strtotime(Request::input('due_end'));
-        } elseif (isset($data['due_start'])) {
-            // If 'due_end' is not provided, set it to 'due_start' value
-            $data['due_end'] = strtotime(Request::input('due_start'));
-        }
-
-        $todo->update($data);
+        $todo->save();
 
         return back()
             ->with('success', 'Todo updated successfully');
@@ -157,6 +122,8 @@ class ProjectTodoController extends Controller
      */
     public function destroy($project_id, Todo $todo): RedirectResponse
     {
+        $this->authorize('delete', [Todo::class, $todo]);
+
         $todo->delete();
 
         return redirect()->route('project.todo.index', $project_id)
